saveDepartments
