+++ apps/meteor/app/livechat/server/lib/LivechatTyped.ts
@@ -16,7 +16,6 @@ import type {
 	TransferData,
 	IOmnichannelAgent,
 	ILivechatDepartmentAgents,
-	LivechatDepartmentDTO,
 	ILivechatInquiryRecord,
 	ILivechatContact,
 	ILivechatContactChannel,
@@ -74,7 +73,6 @@ import { createContact, createContactFromVisitor, isSingleContactEnabled } from 
 import { parseAgentCustomFields, updateDepartmentAgents, validateEmail, normalizeTransferredByData } from './Helper';
 import { QueueManager } from './QueueManager';
 import { RoutingManager } from './RoutingManager';
-import { isDepartmentCreationAvailable } from './isDepartmentCreationAvailable';
 import type { CloseRoomParams, CloseRoomParamsByUser, CloseRoomParamsByVisitor, ILivechatMessage } from './localTypes';
 import { parseTranscriptRequest } from './parseTranscriptRequest';
 
@@ -1773,134 +1771,6 @@ class LivechatClass {
 		void notifyOnRoomChangedById(roomData._id);
 
 		return true;
-	}
-
-	/**
-	 * @param {string|null} _id - The department id
-	 * @param {Partial<import('@rocket.chat/core-typings').ILivechatDepartment>} departmentData
-	 * @param {{upsert?: { agentId: string; count?: number; order?: number; }[], remove?: { agentId: string; count?: number; order?: number; }}} [departmentAgents] - The department agents
-	 * @param {{_id?: string}} [departmentUnit] - The department's unit id
-	 */
-	async saveDepartment(
-		userId: string,
-		_id: string | null,
-		departmentData: LivechatDepartmentDTO,
-		departmentAgents?: {
-			upsert?: { agentId: string; count?: number; order?: number }[];
-			remove?: { agentId: string; count?: number; order?: number };
-		},
-		departmentUnit?: { _id?: string },
-	) {
-		check(_id, Match.Maybe(String));
-		if (departmentUnit?._id !== undefined && typeof departmentUnit._id !== 'string') {
-			throw new Meteor.Error('error-invalid-department-unit', 'Invalid department unit id provided', {
-				method: 'livechat:saveDepartment',
-			});
-		}
-
-		const department = _id
-			? await LivechatDepartment.findOneById(_id, { projection: { _id: 1, archived: 1, enabled: 1, parentId: 1 } })
-			: null;
-
-		if (departmentUnit && !departmentUnit._id && department && department.parentId) {
-			const isLastDepartmentInUnit = (await LivechatDepartment.countDepartmentsInUnit(department.parentId)) === 1;
-			if (isLastDepartmentInUnit) {
-				throw new Meteor.Error('error-unit-cant-be-empty', "The last department in a unit can't be removed", {
-					method: 'livechat:saveDepartment',
-				});
-			}
-		}
-
-		if (!department && !(await isDepartmentCreationAvailable())) {
-			throw new Meteor.Error('error-max-departments-number-reached', 'Maximum number of departments reached', {
-				method: 'livechat:saveDepartment',
-			});
-		}
-
-		if (department?.archived && departmentData.enabled) {
-			throw new Meteor.Error('error-archived-department-cant-be-enabled', 'Archived departments cant be enabled', {
-				method: 'livechat:saveDepartment',
-			});
-		}
-
-		const defaultValidations: Record<string, Match.Matcher<any> | BooleanConstructor | StringConstructor> = {
-			enabled: Boolean,
-			name: String,
-			description: Match.Optional(String),
-			showOnRegistration: Boolean,
-			email: String,
-			showOnOfflineForm: Boolean,
-			requestTagBeforeClosingChat: Match.Optional(Boolean),
-			chatClosingTags: Match.Optional([String]),
-			fallbackForwardDepartment: Match.Optional(String),
-			departmentsAllowedToForward: Match.Optional([String]),
-			allowReceiveForwardOffline: Match.Optional(Boolean),
-		};
-
-		// The Livechat Form department support addition/custom fields, so those fields need to be added before validating
-		Object.keys(departmentData).forEach((field) => {
-			if (!defaultValidations.hasOwnProperty(field)) {
-				defaultValidations[field] = Match.OneOf(String, Match.Integer, Boolean);
-			}
-		});
-
-		check(departmentData, defaultValidations);
-		check(
-			departmentAgents,
-			Match.Maybe({
-				upsert: Match.Maybe(Array),
-				remove: Match.Maybe(Array),
-			}),
-		);
-
-		const { requestTagBeforeClosingChat, chatClosingTags, fallbackForwardDepartment } = departmentData;
-		if (requestTagBeforeClosingChat && (!chatClosingTags || chatClosingTags.length === 0)) {
-			throw new Meteor.Error(
-				'error-validating-department-chat-closing-tags',
-				'At least one closing tag is required when the department requires tag(s) on closing conversations.',
-				{ method: 'livechat:saveDepartment' },
-			);
-		}
-
-		if (_id && !department) {
-			throw new Meteor.Error('error-department-not-found', 'Department not found', {
-				method: 'livechat:saveDepartment',
-			});
-		}
-
-		if (fallbackForwardDepartment === _id) {
-			throw new Meteor.Error(
-				'error-fallback-department-circular',
-				'Cannot save department. Circular reference between fallback department and department',
-			);
-		}
-
-		if (fallbackForwardDepartment) {
-			const fallbackDep = await LivechatDepartment.findOneById(fallbackForwardDepartment, {
-				projection: { _id: 1, fallbackForwardDepartment: 1 },
-			});
-			if (!fallbackDep) {
-				throw new Meteor.Error('error-fallback-department-not-found', 'Fallback department not found', {
-					method: 'livechat:saveDepartment',
-				});
-			}
-		}
-
-		const departmentDB = await LivechatDepartment.createOrUpdateDepartment(_id, departmentData);
-		if (departmentDB && departmentAgents) {
-			await updateDepartmentAgents(departmentDB._id, departmentAgents, departmentDB.enabled);
-		}
-
-		// Disable event
-		if (department?.enabled && !departmentDB?.enabled) {
-			await callbacks.run('livechat.afterDepartmentDisabled', departmentDB);
-		}
-
-		if (departmentUnit) {
-			await callbacks.run('livechat.manageDepartmentUnit', { userId, departmentId: departmentDB._id, unitId: departmentUnit._id });
-		}
-
-		return departmentDB;
 	}
 }
